+++ lib/bali/form_builder/range_fields.rb
@@ -14,7 +14,8 @@ module Bali
         xs: "range-xs",
         sm: "range-sm",
         md: "range-md",
-        lg: "range-lg"
+        lg: "range-lg",
+        xl: "range-xl"
       }.freeze
 
       COLORS = {
@@ -95,7 +96,7 @@ module Bali
         range_class = [
           RANGE_CLASS,
           "w-full",
-          SIZES[options[:size]],
+          size_variant(options, SIZES),
           COLORS[options[:color]],
           (errors?(method, options) ? "range-error" : nil),
           options[:class]